Split fs file into multiple files
